+++ api/wild_encounters/wild_encounter_routes.py
@@ -8,6 +8,7 @@ class WildEncounterRoutes():
    ROUTES: dict[ str, PostRouteHandler ] = {
    '/get-wild-encounters': WildEncounterController.get_wild_encounters,
    '/get-wild-encounter-names': WildEncounterController.get_wild_encounter_names,
+   '/get-wild-encounter-schedule-options': WildEncounterController.get_wild_encounter_schedule_options,
    '/get-wild-encounter-occurrences': WildEncounterController.get_wild_encounter_occurrences,
    '/get-wild-encounter-schedule-times': WildEncounterController.get_wild_encounter_schedule_times,
    '/set-wild-encounter-schedule': WildEncounterController.set_wild_encounter_schedule,